refactor: 메서드 파일 분리
